+++ ChangeLog.md
@@ -12,3 +12,6 @@
 
 ### 2024-05-15
 1. 完成迁移 angular - @angular/core 内容
+
+### 2024-05-15
+1. 完成迁移 angular - @angular/forms 内容